refactor: [#20] Home 컴포넌트 분리 - Home의 컴포넌트들을 각각 분리하여 재사용이 가능하도록 함
